refactor sklearn.pca
